# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Localiza Integração Cenário 4
  description: API demonstrativa para CRM, reservas e contratos, financeiro e faturamento,
    atendimento e assistência 24h, workflow e integração.
  version: 1.0.0
paths:
...
  /api/v1/support/tickets:
    post:
      tags:
      - Atendimento e assistência 24h
      summary: Open Ticket
      description: Abre chamado com SLA do contrato; sem Contracts disponível, fica
        pendente de elegibilidade.
      operationId: open_ticket_api_v1_support_tickets_post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          maxLength: 100
          examples:
          - demo-ticket-001
          title: Idempotency-Key
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '201':
          description: Chamado criado ou resposta idempotente recuperada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
              examples:
                open:
                  summary: Chamado elegível aberto
                  value:
                    ticketId: dddddddd-dddd-4ddd-8ddd-dddddddddddd
                    customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
                    contractId: bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
                    serviceCode: RENTAL-FLEX
                    category: OUTAGE
                    status: OPEN
                    priority: HIGH
                    slaHours: 8
                    dueAt: '2026-10-01T18:10:00Z'
                pending:
                  summary: Chamado pendente por indisponibilidade de Contracts
                  value:
                    ticketId: dddddddd-dddd-4ddd-8ddd-dddddddddddd
                    customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
                    contractId: bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
                    serviceCode: RENTAL-FLEX
                    category: QUESTION
                    status: PENDING_ENTITLEMENT
                    priority: NORMAL
                replayed:
                  summary: Resposta recuperada com Idempotency-Replayed true
                  value:
                    ticketId: dddddddd-dddd-4ddd-8ddd-dddddddddddd
                    customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
                    contractId: bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
                    serviceCode: RENTAL-FLEX
                    category: OUTAGE
                    status: OPEN
                    priority: HIGH
                    slaHours: 8
                    dueAt: '2026-10-01T18:10:00Z'
          headers:
            X-Correlation-ID:
              description: UUID usado para correlacionar a requisição, os logs e os
                eventos.
              schema:
                type: string
                format: uuid
              example: 11111111-1111-4111-8111-111111111111
            Idempotency-Replayed:
              description: Retorna true quando a API reutiliza a resposta armazenada.
              schema:
                type: string
                enum:
                - 'true'
                - 'false'
              example: 'false'
        '400':
          description: Cabeçalho de correlação inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:bad-request
                title: Requisição inválida
                status: 400
                detail: X-Correlation-ID deve ser UUID
                instance: /api/v1/contracts/drafts
                code: BAD_REQUEST
                correlationId: 11111111-1111-4111-8111-111111111111
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/contracts/drafts
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/contracts/drafts
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: Idempotency-Key reutilizada com outro chamado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:idempotency-conflict
                title: Conflito
                status: 409
                detail: Idempotency-Key já utilizada com outro chamado
                instance: /api/v1/support/tickets
                code: IDEMPOTENCY_CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida ou contrato, serviço ou cliente sem elegibilidade.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                businessRule:
                  summary: Contrato sem elegibilidade
                  value:
                    type: urn:archcorp:problem:business-rule-violation
                    title: Entrada inválida
                    status: 422
                    detail: Contrato, serviço ou cliente sem elegibilidade
                    instance: /api/v1/support/tickets
                    code: BUSINESS_RULE_VIOLATION
                    correlationId: 11111111-1111-4111-8111-111111111111
                validation:
                  summary: Corpo da requisição inválido
                  value:
                    type: urn:archcorp:problem:validation-error
                    title: Entrada inválida
                    status: 422
                    detail:
                    - type: uuid_parsing
                      loc:
                      - body
                      - contractId
                      msg: Input should be a valid UUID
                      input: not-a-uuid
                    instance: /api/v1/support/tickets
                    code: VALIDATION_ERROR
                    correlationId: 11111111-1111-4111-8111-111111111111
                    errors:
                    - type: uuid_parsing
                      loc:
                      - body
                      - contractId
                      msg: Input should be a valid UUID
                      input: not-a-uuid
    get:
      tags:
      - Atendimento e assistência 24h
      summary: List Tickets
      description: Lista até 200 chamados.
      operationId: list_tickets_api_v1_support_tickets_get
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketDetailResponse'
                title: Response List Tickets Api V1 Support Tickets Get
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/support/tickets
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/support/tickets
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
      parameters:
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
...
  /api/v1/support/tickets/{ticket_id}:
    get:
      tags:
      - Atendimento e assistência 24h
      summary: Get Ticket
      description: Consulta um chamado.
      operationId: get_ticket_api_v1_support_tickets__ticket_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Ticket Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/support/tickets/{ticket_id}
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/support/tickets/{ticket_id}
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Chamado não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Chamado não encontrado
                instance: /api/v1/support/tickets/{ticket_id}
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/support/tickets/{ticket_id}
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
  /api/v1/support/tickets/{ticket_id}/assign:
    post:
      tags:
      - Atendimento e assistência 24h
      summary: Assign Ticket
      description: Define o responsável de um chamado aberto ou pendente.
      operationId: assign_ticket_api_v1_support_tickets__ticket_id__assign_post
      security:
      - HTTPBearer: []
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Ticket Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/support/tickets/{ticket_id}/assign
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/support/tickets/{ticket_id}/assign
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Chamado não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Chamado não encontrado
                instance: /api/v1/support/tickets/{ticket_id}/assign
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: Chamado resolvido ou rejeitado não pode ser atribuído.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: Chamado resolvido ou rejeitado não pode ser atribuído
                instance: /api/v1/support/tickets/{ticket_id}/assign
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/support/tickets/{ticket_id}/assign
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
  /api/v1/support/tickets/{ticket_id}/resolve:
    post:
      tags:
      - Atendimento e assistência 24h
      summary: Resolve Ticket
      description: Resolve o chamado aberto e publica TicketResolved.v1.
      operationId: resolve_ticket_api_v1_support_tickets__ticket_id__resolve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Ticket Id
      - name: X-Correlation-ID
        in: header
        required: false
        description: UUID de correlação. A API gera um UUID quando o cabeçalho é omitido.
        schema:
          type: string
          format: uuid
        example: 11111111-1111-4111-8111-111111111111
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          description: Token ausente ou inválido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:unauthorized
                title: Não autenticado
                status: 401
                detail: Token ausente ou inválido
                instance: /api/v1/support/tickets/{ticket_id}/resolve
                code: UNAUTHORIZED
                correlationId: 11111111-1111-4111-8111-111111111111
        '403':
          description: Papel sem permissão para a operação.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:forbidden
                title: Acesso negado
                status: 403
                detail: Papel sem permissão para esta operação
                instance: /api/v1/support/tickets/{ticket_id}/resolve
                code: FORBIDDEN
                correlationId: 11111111-1111-4111-8111-111111111111
        '404':
          description: Chamado não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:not-found
                title: Recurso não encontrado
                status: 404
                detail: Chamado não encontrado
                instance: /api/v1/support/tickets/{ticket_id}/resolve
                code: NOT_FOUND
                correlationId: 11111111-1111-4111-8111-111111111111
        '409':
          description: Somente chamado aberto pode ser resolvido.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:conflict
                title: Conflito
                status: 409
                detail: Somente chamado aberto pode ser resolvido
                instance: /api/v1/support/tickets/{ticket_id}/resolve
                code: CONFLICT
                correlationId: 11111111-1111-4111-8111-111111111111
        '422':
          description: Entrada inválida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: urn:archcorp:problem:validation-error
                title: Entrada inválida
                status: 422
                detail:
                - *id001
                instance: /api/v1/support/tickets/{ticket_id}/resolve
                code: VALIDATION_ERROR
                correlationId: 11111111-1111-4111-8111-111111111111
                errors:
                - *id001
...
components:
  schemas:
    AssignmentInput:
      properties:
        owner:
          type: string
          maxLength: 80
          minLength: 2
          title: Owner
      type: object
      required:
      - owner
      title: AssignmentInput
      examples:
      - owner: Equipe de assistência
...
    TicketCreate:
      properties:
        customerId:
          type: string
          format: uuid
          title: Customerid
        contractId:
          type: string
          format: uuid
          title: Contractid
        serviceCode:
          type: string
          title: Servicecode
        category:
          type: string
          maxLength: 80
          minLength: 2
          title: Category
        description:
          type: string
          maxLength: 2000
          minLength: 3
          title: Description
      type: object
      required:
      - customerId
      - contractId
      - serviceCode
      - category
      - description
      title: TicketCreate
      examples:
      - category: OUTAGE
        contractId: bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
        customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
        description: Serviço indisponível durante a demonstração
        serviceCode: RENTAL-FLEX
    TicketDetailResponse:
      properties:
        ticketId:
          type: string
          format: uuid
          title: Ticketid
        customerId:
          type: string
          format: uuid
          title: Customerid
        contractId:
          type: string
          format: uuid
          title: Contractid
        serviceCode:
          type: string
          title: Servicecode
        category:
          type: string
          title: Category
        status:
          type: string
          enum:
          - OPEN
          - PENDING_ENTITLEMENT
          - REJECTED_ENTITLEMENT
          - RESOLVED
          title: Status
        priority:
          type: string
          enum:
          - NORMAL
          - HIGH
          title: Priority
        slaHours:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slahours
        dueAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Dueat
        description:
          type: string
          title: Description
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
      type: object
      required:
      - ticketId
      - customerId
      - contractId
      - serviceCode
      - category
      - status
      - priority
      - slaHours
      - dueAt
      - description
      - owner
      title: TicketDetailResponse
      examples:
      - category: OUTAGE
        contractId: bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
        customerId: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
        description: Serviço indisponível durante a demonstração
        dueAt: '2026-10-01T18:10:00Z'
        owner: Equipe de assistência
        priority: HIGH
        serviceCode: RENTAL-FLEX
        slaHours: 8
        status: OPEN
        ticketId: dddddddd-dddd-4ddd-8ddd-dddddddddddd
    TicketResponse:
      properties:
        ticketId:
          type: string
          format: uuid
          title: Ticketid
        customerId:
          type: string
          format: uuid
          title: Customerid
        contractId:
          type: string
          format: uuid
          title: Contractid
        serviceCode:
          type: string
          title: Servicecode
        category:
          type: string
          title: Category
        status:
          type: string
          enum:
          - OPEN
          - PENDING_ENTITLEMENT
          - REJECTED_ENTITLEMENT
          title: Status
        priority:
          type: string
          enum:
          - NORMAL
          - HIGH
          title: Priority
        slaHours:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slahours
        dueAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Dueat
      type: object
      required:
      - ticketId
      - customerId
      - contractId
      - serviceCode
      - category
      - status
      - priority
      - slaHours
      - dueAt
      title: TicketResponse
    ProblemDetails:
      description: Corpo de erro publicado no OpenAPI.
      examples:
      - code: NOT_FOUND
        correlationId: 11111111-1111-4111-8111-111111111111
        detail: Contrato não encontrado
        instance: /api/v1/contracts/bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
        status: 404
        title: Recurso não encontrado
        type: urn:archcorp:problem:not-found
      properties:
        type:
          description: URI que identifica o tipo do problema.
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          anyOf:
          - type: string
          - items:
              additionalProperties: true
              type: object
            type: array
          description: Mensagem legível. Em 422 mantém a lista de erros da versão
            1.
          title: Detail
        instance:
          description: Caminho da requisição que originou o erro.
          title: Instance
          type: string
        code:
          description: Código estável para tratamento pelo cliente.
          title: Code
          type: string
        correlationId:
          title: Correlationid
          type: string
        errors:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          default: null
          title: Errors
      required:
      - type
      - title
      - status
      - detail
      - instance
      - code
      - correlationId
      title: ProblemDetails
      type: object
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer